refactor: 패키지를 역할 별로 나눈다
