+++ src/main/java/study/calculator/InputNumber.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import java.util.Objects;
