+++ src/main/java/study/calculator/Operator.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import java.util.function.BiFunction;
 import java.util.stream.Stream;
