+++ src/main/java/study/racinggame/RacingCarApplication.java
@@ -1,4 +1,6 @@
-package study;
+package study.racinggame;
+
+import study.racinggame.controller.Game;
 
 public class RacingCarApplication {
     public static void main(String[] args) {
