+++ src/main/java/study/racinggame/domain/RandomNumber.java
@@ -1,4 +1,4 @@
-package study;
+package study.racinggame.domain;
 
 import java.util.Random;
